+++ Modules/Microsoft365DSC/DSCResources/MSFT_AADAppManagementPolicy/MSFT_AADAppManagementPolicy.psm1
@@ -376,6 +376,10 @@ function Export-TargetResource
     param
     (
         [Parameter()]
+        [System.String]
+        $Filter,
+
+        [Parameter()]
         [System.Management.Automation.PSCredential]
         $Credential,
 
@@ -422,7 +426,7 @@ function Export-TargetResource
     try
     {
         $Script:ExportMode = $true
-        [array] $Script:exportedInstances = Get-MgBetaPolicyAppManagementPolicy -ErrorAction Stop
+        [array] $Script:exportedInstances = Get-MgBetaPolicyAppManagementPolicy -Filter $Filter -All -ErrorAction Stop
 
         $i = 1
         $dscContent = ''