+++ Modules/Microsoft365DSC/DSCResources/MSFT_AADEntitlementManagementAccessPackageAssignmentPolicy/MSFT_AADEntitlementManagementAccessPackageAssignmentPolicy.psm1
@@ -754,6 +754,10 @@ function Export-TargetResource
     param
     (
         [Parameter()]
+        [System.String]
+        $Filter,
+
+        [Parameter()]
         [System.Management.Automation.PSCredential]
         $Credential,
 
@@ -801,6 +805,7 @@ function Export-TargetResource
     {
         [array]$getValue = Get-MgBetaEntitlementManagementAccessPackageAssignmentPolicy `
             -All `
+            -Filter $Filter `
             -ErrorAction Stop
 
         $i = 1