+++ Modules/Microsoft365DSC/DSCResources/MSFT_TeamsCortanaPolicy/MSFT_TeamsCortanaPolicy.psm1
@@ -265,6 +265,10 @@ function Export-TargetResource
     param
     (
         [Parameter()]
+        [System.String]
+        $Filter = "*",
+
+        [Parameter()]
         [System.Management.Automation.PSCredential]
         $Credential,
 
@@ -310,7 +314,7 @@ function Export-TargetResource
 
     try
     {
-        [array]$getValue = Get-CsTeamsCortanaPolicy -ErrorAction Stop
+        [array]$getValue = Get-CsTeamsCortanaPolicy -Filter $Filter -ErrorAction Stop
 
         $i = 1
         $dscContent = ''