+++ Modules/Microsoft365DSC/DSCResources/MSFT_TeamsNetworkRoamingPolicy/MSFT_TeamsNetworkRoamingPolicy.psm1
@@ -274,6 +274,10 @@ function Export-TargetResource
     param
     (
         [Parameter()]
+        [System.String]
+        $Filter = "*",
+
+        [Parameter()]
         [System.Management.Automation.PSCredential]
         $Credential,
 
@@ -319,7 +323,7 @@ function Export-TargetResource
 
     try
     {
-        [array]$getValue = Get-CsTeamsNetworkRoamingPolicy -ErrorAction Stop
+        [array]$getValue = Get-CsTeamsNetworkRoamingPolicy -Filter $Filter -ErrorAction Stop
 
         $i = 1
         $dscContent = ''