+++ Modules/Microsoft365DSC/DSCResources/MSFT_TeamsOnlineVoicemailUserSettings/MSFT_TeamsOnlineVoicemailUserSettings.psm1
@@ -363,6 +363,10 @@ function Export-TargetResource
     param
     (
         [Parameter()]
+        [System.String]
+        $Filter,
+
+        [Parameter()]
         [System.Management.Automation.PSCredential]
         $Credential,
 
@@ -404,7 +408,14 @@ function Export-TargetResource
 
     try
     {
-        $allUsers = Get-CsOnlineUser -Properties 'UserPrincipalName'
+        $splat = @{
+            Properties = 'UserPrincipalName'
+        }
+        if (-not [System.String]::IsNullOrEmpty($Filter))
+        {
+            $splat.Add('Filter', $Filter)
+        }
+        $allUsers = Get-CsOnlineUser @splat
         $i = 1
         Write-M365DSCHost -Message "`r`n" -DeferWrite
         $dscContent = [System.Text.StringBuilder]::new()