+++ Modules/Microsoft365DSC/DSCResources/MSFT_TeamsUnassignedNumberTreatment/MSFT_TeamsUnassignedNumberTreatment.psm1
@@ -304,6 +304,10 @@ function Export-TargetResource
     param
     (
         [Parameter()]
+        [System.String]
+        $Filter = "*",
+
+        [Parameter()]
         [System.Management.Automation.PSCredential]
         $Credential,
 
@@ -349,7 +353,7 @@ function Export-TargetResource
 
     try
     {
-        [array]$getValue = Get-CsTeamsUnassignedNumberTreatment -ErrorAction Stop
+        [array]$getValue = Get-CsTeamsUnassignedNumberTreatment -Filter $Filter -ErrorAction Stop
 
         $i = 1
         $dscContent = ''